rename layouts to be more a meaninful concept around what it represents
