+++ internal/bubbles/layout/xpnavigator/events.go
@@ -1,4 +1,4 @@
-package navigator
+package xpnavigator
 
 import (
 	"time"
