+++ internal/bubbles/layout/xpsummary/model.go
@@ -1,4 +1,4 @@
-package viewer
+package xpsummary
 
 import (
 	"fmt"
